Initial on the wire exploration
This work is from time spent debugging Chrome DevTools' actual activity. I've identified what I
believe to be the main surface that needs coverage to get a basic REPL working. Atm just trying
to get something really simple talking, and then can go back in and correct things like
marshalling, better scripts/frames/urls/objectIds/&c. The current vm is bare- it needs more of the
module wrapper given to it.

The architecture needs some more careful consideration, at least if multiple targets is a desired
goal. Currently DevTools seems to assume it's going to open a connection and the DevTools Protocol
server on the other end will attach it to the appropriate target, and I don't think it will change
that target? Maybe? I haven't been testing with complex pages that have SW or frames to really see.
It'd be great if this library really afforded users the ability to build as many servers as they
desired, each one with whatever collection of targets they wanted, and could handle clients that
did things like attach to multiple targets, or detach from some of it's targets. This chiefly
implies that a connection is not at all strongly owned by a target- there's some kind of server
top level thing that handles connections & dispatches out to it's set of connections.

Should be pretty close to having some basic wire convos though, & that's cool.