Track RemoteObject & their objectId in connection.
This starts the work of having devrep keep track of objects, which it has to
do to allow the ongoing chatter with the debugger. Using the console returns
a preview object, and an objectId, that the server has to keep track of and
be able to use to answer followup questions getting more information about
the remote object.

By scoping objectId's to connections, we:
1. Create a security barrier, since only your connections objects can be looked up
2. Have a connection scope, such that when the connection goes away all objects can finally be freed.